+++ packages/next-common/components/overview/activeProposals/advisoryCommittee.js
@@ -2,10 +2,7 @@ import {
   getAdvisoryCommitteeMenu,
   Names,
 } from "next-common/utils/consts/menu/advisoryCouncil";
-import {
-  getProposalPostTitleColumn,
-  getStatusTagColumn,
-} from "./columns/common";
+import { getProposalPostTitleColumn, getStatusTagColumn } from "./columns";
 import businessCategory from "next-common/utils/consts/business/category";
 import { overviewApi } from "next-common/services/url";
 import { toAdvisoryMotionsListItem } from "next-common/utils/viewfuncs";
