+++ packages/next-common/components/overview/activeProposals/alliance.js
@@ -1,8 +1,5 @@
 import { getAllianceMenu, Names } from "next-common/utils/consts/menu/alliance";
-import {
-  getProposalPostTitleColumn,
-  getStatusTagColumn,
-} from "./columns/common";
+import { getProposalPostTitleColumn, getStatusTagColumn } from "./columns";
 import normalizeAllianceMotion from "next-common/utils/viewfuncs/alliance/allianceMotion";
 import normalizeAllianceAnnouncement from "next-common/utils/viewfuncs/alliance/allianceAnnouncement";
 import businessCategory from "next-common/utils/consts/business/category";
