+++ packages/next-common/components/overview/activeProposals/council.js
@@ -1,9 +1,6 @@
 import { CHAIN } from "next-common/utils/constants";
 import { getCouncilMenu, Names } from "next-common/utils/consts/menu/council";
-import {
-  getProposalPostTitleColumn,
-  getStatusTagColumn,
-} from "./columns/common";
+import { getProposalPostTitleColumn, getStatusTagColumn } from "./columns";
 import businessCategory from "next-common/utils/consts/business/category";
 import { overviewApi } from "next-common/services/url";
 import normalizeCouncilMotionListItem from "next-common/utils/viewfuncs/collective/normalizeCouncilMotionListItem";
